// packages/studio-ui/src/types.ts
export interface DefaultNodeProps {
  [prop: string]: unknown;
}
// ...
export interface StudioComponentDefinition<P = DefaultNodeProps> {
  props: PropDefinitions<P>;
  module: string;
  importedName: string;
}